requirements.txt
attrs
fire
tqdm
posix_ipc
